test/fixtures/wpt/webidl/ecmascript-binding/legacy-platform-object/OwnPropertyKeys.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Legacy platform objects [[OwnPropertyKeys]] method</title>
<link rel="help" href="https://webidl.spec.whatwg.org/#legacy-platform-object-ownpropertykeys">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    function custom_assert_array_equals(actual, expected, msg) {
        function replacer(k, v) {
          if (typeof v === "symbol") {
            return v.toString();
          }
          return v;
        }
        assert_array_equals(actual, expected, " actual " + JSON.stringify(actual, replacer) + " expected " + JSON.stringify(expected, replacer));
    }

    test(function() {
        var element = document.createElement("div");
        element.appendChild(document.createElement("div"));
        element.appendChild(document.createElement("div"));
        element.appendChild(document.createElement("div"));
        custom_assert_array_equals(Reflect.ownKeys(element.childNodes), ["0", "1", "2"]);
    }, "must enumerate property indices in ascending numerical order");

    test(function() {
        var element = document.createElement("div");
        element.setAttribute("data-foo", "foo content");
        element.setAttribute("data-bar", "bar content");
        element.setAttribute("data-baz", "baz content");
        custom_assert_array_equals(Reflect.ownKeys(element.dataset), ["foo", "bar", "baz"]);
    }, "must enumerate property names in list order");


    test(function() {
        var element = document.createElement("div");
        element.setAttribute("id", "foo");
        element.setAttribute("class", "bar");
        custom_assert_array_equals(Reflect.ownKeys(element.attributes), ["0", "1", "id", "class"]);
    }, "must first enumerate property indices in ascending numerical order, then named properties in list order");


    test(function() {
        var element = document.createElement("div");
        element.attributes.foo = "some value";
        element.attributes.bar = "and another";
        element.setAttribute("id", "foo");
        element.setAttribute("class", "bar");
        custom_assert_array_equals(Reflect.ownKeys(element.attributes), ["0", "1", "id", "class", "foo", "bar"]);
    }, "must enumerate own properties after indexed and named properties even when they're added first");

    test(function() {
        var symb1 = Symbol();
        var symb2 = Symbol();
        var element = document.createElement("div");
        element.attributes.foo = "some value";
        element.attributes[symb1] = "symb1";
        element.attributes[symb2] = "symb2";
        element.attributes.bar = "and another";
        element.setAttribute("id", "foo");
        element.setAttribute("class", "bar");
        custom_assert_array_equals(Reflect.ownKeys(element.attributes),
            ["0", "1", "id", "class", "foo", "bar", symb1, symb2]);
    }, "must enumerate symbols after strings, regardless of which ones got added first");
</script>
